requirements.txt
# MLB Daily AI Creator

# Common dependencies
python-dotenv>=1.0.0
supabase>=2.0.0  # For Phase 2 & 4

# Phase 2: Streamlit Dashboard
streamlit>=1.40.0

# Phase 3: Video Pipeline
edge-tts>=7.0
moviepy>=2.0
Pillow>=10.0
requests>=2.30

# Phase 4: YouTube Upload
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
